requirements.txt
python-telegram-bot==21.3
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
notion-client==2.2.1
aiohttp==3.9.5
